Chart Refactoring
* Move the chart to the charts folder
* Use helm-docs to generate docs for the chart
* Use chart release action to automate chart releasing process
* Delete deploy folder to only maintain a single installation option (chart)

Other changes:
* Add Makefile
* Bump min go version to 1.21
* Bump golangci-lint 1.55.2
