+++ .golangci.yml
@@ -1,0 +1,10 @@
+run:
+  timeout: 10m
+
+linters:
+  disable-all: true
+  enable:
+    - govet
+    - revive
+    - promlinter
+    - gofmt
