+++ charts/cr.yaml
@@ -1,0 +1,2 @@
+owner: deckhouse
+git-base-url: https://api.github.com/
