+++ charts/k8s-image-availability-exporter/Chart.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+appVersion: "v0.5.1"
+description: Application for monitoring the cluster workloads image presence in a container registry.
+name: k8s-image-availability-exporter
+version: "0.6.0"
+kubeVersion: ">=1.14.0-0"
