+++ charts/k8s-image-availability-exporter/README.md.gotmpl
@@ -1,0 +1,31 @@
+{{ template "chart.header" . }}
+
+{{ template "chart.badgesSection" . }}
+
+{{ template "chart.description" . }}
+
+{{ template "chart.requirementsSection" . }}
+
+{{ template "chart.deprecationWarning" . }}
+
+## Introduction
+
+This chart bootstraps a [k8s-image-availability-exporter](https://github.com/flant/k8s-image-availability-exporter) deployment on a [Kubernetes](http://kubernetes.io) cluster using the [Helm](https://helm.sh) package manager.
+
+{{ template "chart.homepageLine" . }}
+
+{{ template "chart.maintainersSection" . }}
+
+{{ template "chart.valuesSection" . }}
+
+Specify each parameter using the `--set key=value[,key=value]` argument to `helm install`. For example,
+
+```bash
+helm install my-release k8s-image-availability-exporter --set k8sImageAvailabilityExporter.replicas=2
+```
+
+Alternatively, one or more YAML files that specify the values for the above parameters can be provided while installing the chart. For example,
+
+```bash
+helm install my-release k8s-image-availability-exporter -f values1.yaml,values2.yaml
+```
