+++ charts/k8s-image-availability-exporter/templates/_helpers.tpl
