+++ charts/k8s-image-availability-exporter/templates/prometheus-rule.yaml
