+++ charts/k8s-image-availability-exporter/templates/rbac.yaml
