+++ charts/k8s-image-availability-exporter/templates/service-monitor.yaml
