--- deploy/deployment.yaml
@@ -1,47 +1,0 @@
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: image-availability-exporter
-  namespace: kube-system
----
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: image-availability-exporter
-  namespace: kube-system
-spec:
-  selector:
-    matchLabels:
-      app: image-availability-exporter
-  strategy:
-    type: Recreate
-  template:
-    metadata:
-      labels:
-        app: image-availability-exporter
-    spec:
-      containers:
-        - args:
-            - --bind-address=:8080
-          ports:
-            - containerPort: 8080
-              name: http-metrics
-          image: registry-write.deckhouse.io/k8s-image-availability-exporter/k8s-image-availability-exporter:v0.3.0
-          imagePullPolicy: IfNotPresent
-          livenessProbe:
-            httpGet:
-              path: /healthz
-              port: 8080
-              scheme: HTTPS
-          name: image-availability-exporter
-          readinessProbe:
-            httpGet:
-              path: /healthz
-              port: 8080
-              scheme: HTTP
-          resources:
-            requests:
-              cpu: 100m
-              memory: 64Mi
-      serviceAccountName: image-availability-exporter
