--- deploy/rbac.yaml
@@ -1,52 +1,0 @@
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRole
-metadata:
-  name: image-availability-exporter
-rules:
-  - apiGroups:
-      - ""
-    resources:
-      - secrets
-    verbs:
-      - list
-      - watch
-      - get
-  - apiGroups:
-      - ""
-    resources:
-      - serviceaccounts
-    verbs:
-      - get
-  - apiGroups:
-      - extensions
-      - apps
-    resources:
-      - deployments
-      - daemonsets
-      - statefulsets
-    verbs:
-      - list
-      - watch
-      - get
-  - apiGroups:
-      - batch
-    resources:
-      - cronjobs
-    verbs:
-      - list
-      - watch
-      - get
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: ClusterRoleBinding
-metadata:
-  name: image-availability-exporter
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: image-availability-exporter
-subjects:
-  - kind: ServiceAccount
-    name: image-availability-exporter
-    namespace: kube-system
