--- helm/charts/k8s-image-availability-exporter/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v1
-appVersion: "v0.5.1"
-description: Application for monitoring the cluster workloads image presence in docker registry.
-name: k8s-image-availability-exporter
-version: "0.5.1"
