--- helm/charts/k8s-image-availability-exporter/values.yaml
@@ -1,18 +1,0 @@
-k8sImageAvailabilityExporter:
-  image:
-    repository: registry.deckhouse.io/k8s-image-availability-exporter/k8s-image-availability-exporter
-    tag: v0.5.1
-    pullPolicy: IfNotPresent
-  replicas: 1
-  resources: {}
-  args:
-    - --bind-address=:8080
-  
-serviceMonitor:
-  enabled: false
-  interval: 15s
-
-prometheusRule:
-  enabled: false
-  defaultGroupsEnabled: true
-  additionalGroups: []
