+++ pkg/registry/image_pull.go
@@ -1,4 +1,4 @@
-package registry_checker
+package registry
 
 import (
 	"errors"
